adding main firebase integration
