+++ firebase.json
@@ -1,0 +1,1 @@
+{"flutter":{"platforms":{"android":{"default":{"projectId":"e-bike-connect-os","appId":"1:618108623420:android:30c7f9cccb9e142c3a4139","fileOutput":"android/app/google-services.json"}},"dart":{"lib/firebase_options.dart":{"projectId":"e-bike-connect-os","configurations":{"android":"1:618108623420:android:30c7f9cccb9e142c3a4139","ios":"1:618108623420:ios:863e1a1876021d123a4139","web":"1:618108623420:web:612015f0603dc9c33a4139"}}}}}}
