+++ pubspec.yaml
@@ -10,6 +10,7 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.6
+  firebase_core: ^3.4.0
 
 dev_dependencies:
   flutter_test:
